% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{add_env_data}
\alias{add_env_data}
\title{Add environmental data to BBS data frame}
\usage{
add_env_data(bbs_data)
}
\arguments{
\item{bbs_data}{dataframe that contains BBS site_id and year columns}
}
\value{
dataframe with original data and associated environmental data
}
\description{
Add environmental data to BBS data frame
}
