% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{add_observers}
\alias{add_observers}
\title{Add observer ID's to data}
\usage{
add_observers(bbs_data)
}
\arguments{
\item{bbs_data}{dataframe that contains BBS site_id and year columns}
}
\description{
Add observer ID's to data
}

