% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{add_ranefs}
\alias{add_ranefs}
\title{Add point estimates for random effects associated with site_id and observer_id}
\usage{
add_ranefs(df, last_training_year)
}
\arguments{
\item{df}{a data.frame, as produced by \link{get_bbs_data}}

\item{last_training_year}{the final year to include in the training set for
fitting the mixed model}
}
\description{
Add point estimates for random effects associated with site_id and observer_id
}

