% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{cleanup_multi_ts_forecasts}
\alias{cleanup_multi_ts_forecasts}
\title{Cleanup output from multi-time-series forecasting into a simple data frame}
\usage{
cleanup_multi_ts_forecasts(ts_forecast_df, groups)
}
\arguments{
\item{ts_forecast_df}{data.frame containing year, site, cast_naive, cast_avg,
and cast_arima columns}
}
\value{
data.frame
}
\description{
Takes the output of a dplyr based run of forecasting on multiple groups and
makes it into a simple data frame for further analysis. Currently this is
specific to the form of data produce insdie of get_ts_forecasts.
}

