% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{cleanup_ts_forecasts}
\alias{cleanup_ts_forecasts}
\title{Cleanup time-series forecast output}
\usage{
cleanup_ts_forecasts(tsmodel_forecasts)
}
\arguments{
\item{ts_model_forecasts}{dataframe with site_id, timeperiod, cast_naive,
cast_arima, cast_exog_arima, and test_set columns. Each cast_ column
contains the full forecast object for each year of the forecast}
}
\value{
data.frame
}
\description{
Transforms a data frame with one row per for each site and columns
with cells containing full forecast output a number of different forecasts,
into one where each row contains the point forecast and forecast intervals
for a single site, time period, model combination.
}

