% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{filter_species}
\alias{filter_species}
\title{Filter poorly sampled BBS species}
\usage{
filter_species(df)
}
\arguments{
\item{df}{dataframe containing an species_id column}
}
\value{
dataframe, filtered version of initial dataframe
}
\description{
Removes waterbirds, shorebirds, owls, kingfishers, knightjars,
dippers. These species are poorly sampled due to their aquatic or
noctural nature. Also removes taxa that were either partially unidentified
(e.g. "sp.") or were considered hybrids (e.g. "A x B") or were listed as more
than one species (e.g. "B / B")
}

