% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer_model.R
\name{fit_observer_model}
\alias{fit_observer_model}
\title{Estimate observer effects}
\usage{
fit_observer_model(stan_file = "mixed.stan", seed = 1,
  output_file = "observer_model.rds")
}
\description{
By default, \code{stan_file} points to a linear mixed model
}

