% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{get_bbs_data}
\alias{get_bbs_data}
\title{Get the primary bbs data file which compiles the counts, route info, and weather
data. Install it via ecodataretriever if needed.}
\usage{
get_bbs_data()
}
\description{
Get the primary bbs data file which compiles the counts, route info, and weather
data. Install it via ecodataretriever if needed.
}

