% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elev_data.R
\name{get_elev_data}
\alias{get_elev_data}
\title{Get Elevation data}
\usage{
get_elev_data()
}
\value{
a data frame including site_id and elevs columns
}
\description{
If elevation data already exists, do nothing, otherwise use the raster
package to download the average elevation in a 40km radius of each BBS route
and push to the SQLite database
}
