% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{get_modern_contig_ts}
\alias{get_modern_contig_ts}
\title{Get data for all contiguous samples ending in a given year}
\usage{
get_modern_contig_ts(df, stopyear)
}
\arguments{
\item{df}{data.frame including a year column}

\item{stopyear}{year of the most recent sample}
}
\value{
Contiguous time-series data frame (data.frame).
A version of the original data frame only including years that
are in a contiguous series with stopyear.
}
\description{
Get data for all contiguous samples ending in a given year.
If only one year of contiguous data are available keep that year.
}
