% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{get_pop_ts_env_data}
\alias{get_pop_ts_env_data}
\title{Get BBS population time-series data with environmental variables}
\usage{
get_pop_ts_env_data(start_yr, end_yr, last_train_year, min_year_percentage)
}
\arguments{
\item{start_yr}{num first year of time-series}

\item{end_yr}{num last year of time-series}

\item{last_train_year}{num last year of the time-series' training portion}

\item{min_year_percentage}{num percent of years of non-missing data between 
start_yr & last_train_year}
}
\value{
dataframe with site_id, lat, long, year, species_id, and abundance
}
\description{
Selects data from the subset of sites that were visited in at least 
`min_year_percentage` percent of years in the training set, and from
all years between `start_yr` to `end_yr`.
}
\details{
Also attaches associated environmental data
}
