% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{get_popdyn_data_w_zeros}
\alias{get_popdyn_data_w_zeros}
\title{Add zeros to population dynamics data}
\usage{
get_popdyn_data_w_zeros(commdyn_data, start_year, stop_year)
}
\arguments{
\item{df}{data.frame including site_id, year, species_id, and abundance columns}
}
\value{
Data frame with real zeros added
}
\description{
For population dynamics data that lacks real zeros, add them
}

