% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{get_richness_ts_env_data}
\alias{get_richness_ts_env_data}
\title{Get BBS richness time-series data with environmental variables}
\usage{
get_richness_ts_env_data(start_yr, end_yr, min_num_yrs)
}
\arguments{
\item{start_yr}{num first year of time-series}

\item{end_yr}{num last year of time-series}

\item{min_num_yrs}{num minimum number of years of data between start_yr & end_yr}
}
\value{
dataframe with site_id, year, and richness
}
\description{
Get BBS richness time-series data with environmental variables
}

