% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{get_route_data}
\alias{get_route_data}
\title{Get route locations in a SpatialPointsDataFrame}
\usage{
get_route_data()
}
\arguments{
\item{projection}{string projection for route data}
}
\value{
a spatial data frame including site_id, long, and lat
}
\description{
Get route locations in a SpatialPointsDataFrame
}
