% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{get_ts_forecasts}
\alias{get_ts_forecasts}
\title{Perform a suite of time-series only forecasts}
\usage{
get_ts_forecasts(grouped_tsdata, timecol, responsecol, exogcols, lag = 1,
  pred_int_levels = c(80, 95))
}
\arguments{
\item{grouped_tsdata}{data.frame is a grouped data frame where the group
indicates how to split the data frame for running time-series models on each
group}

\item{timecol}{str name of the column that includes the time variable}

\item{responsecol}{str name of the column that includes the response variable}

\item{lag}{numeric how many time-steps to forecast and also how many
time-steps in the current time-series to ignore when fitting.}
}
\value{
data.frame

TODO: separate time-steps to hold out from fitting from time-steps to
      forecast, since when not hindcasting these may be different
}
\description{
Performs time-series only forecasting using naive, average, and ARIMA models
for each group in a grouped data frame of time-series.
}

