% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{install_dataset}
\alias{install_dataset}
\title{Install a particular dataset via Ecodata Retriever}
\usage{
install_dataset(dataset)
}
\arguments{
\item{dataset}{name}
}
\description{
Install a particular dataset via Ecodata Retriever
}

