% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{restruct_ts_forecasts}
\alias{restruct_ts_forecasts}
\title{Split time-series forecasts into point forecast and prediction interval dfs}
\usage{
restruct_ts_forecasts(cleaned_ts_model_fcasts)
}
\arguments{
\item{cleaned_ts_model_forecasts}{dataframe with site_id, model, timeperiod,
obs, pt_fcast, and Lo and Hi columns for each level of prediction interval}
}
\value{
list of data.frames. The first value is the pt_fcast part of the
  input data frame. The second value is a long version of the full data.frame
  with columns for the level and interval characterizing each prediction
  interval
}
\description{
Split time-series forecasts into point forecast and prediction interval dfs
}
