% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_provenance.R
\name{save_provenance}
\alias{save_provenance}
\title{Save some information about where data objects came from}
\usage{
save_provenance(object)
}
\description{
Save some information about where data objects came from
}

