% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{viz_forecast}
\alias{viz_forecast}
\title{Visualize a forecast}
\usage{
viz_forecast(data, forecasts, focal_site)
}
\description{
Visualizes a forecast for a single focal site from a group of forecasts
}
