% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer_model.R
\name{fit_observer_model}
\alias{fit_observer_model}
\title{Estimate observer effects}
\usage{
fit_observer_model(stan_file = "mixed.stan", seed = 1,
  output_file = "observer_model.rds")
}
\value{
\code{NULL}. Saves a list of relevant values in `output_file`
  and the `stanfit` object to disk.
}
\description{
The default `mixed.stan` file is a linear mixed model for richness
using site_id and observer_id as random effects.
}

