% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast-bbs-core.R
\name{get_pop_ts_env_data}
\alias{get_pop_ts_env_data}
\title{Get BBS population time-series data with environmental variables}
\usage{
get_pop_ts_env_data(start_yr, end_yr, min_num_yrs)
}
\arguments{
\item{start_yr}{num first year of time-series}

\item{end_yr}{num last year of time-series}

\item{min_num_yrs}{num minimum number of years of data between start_yr & end_yr}
}
\value{
dataframe with site_id, lat, long, year, species_id, and abundance
}
\description{
Selects sites with data spanning 1982 through 2013 containing at least 25
samples during that period.
}
\details{
Attaches associated environmental data
}

